backend: update profile & add new reservation

+++ backend/app/admin/router.py
@@ -1,27 +1,29 @@
 from http import HTTPStatus
-from uuid import UUID
-from typing import Any, List, Optional
+from typing import Any, Optional
 
 from fastapi import Body, Path, Query, Depends, Request, APIRouter, HTTPException
 
 from app.auth.models import UserModel
-from app.core.security import get_current_user_authorizer
-from app.core.services import create_aliased_response
-from app.property.models import (
+from app.admin.models import (
     PropertyType,
     PropertyInCreate,
     PropertyInUpdate,
+    ReservationModel,
     PropertyInResponse,
     PropertyFilterParams,
     ManyPropertiesInResponse,
 )
-from app.property.services import (
+from app.core.security import get_current_user_authorizer
+from app.core.services import create_aliased_response
+from app.admin.services import (
     create_property,
+    create_reservation,
     check_property_owner,
+    check_user_permission,
     delete_property_by_slug,
     update_property_by_slug,
 )
-from app.property.selectors import (
+from app.admin.selectors import (
     get_properties,
     get_user_properties,
     get_property_by_slug,
@@ -34,6 +36,7 @@ router = APIRouter(prefix="/properties")
     "/feed",
     response_model=ManyPropertiesInResponse,
     status_code=HTTPStatus.OK,
+    tags=["Properties"],
 )
 async def get_properties_feed(
     request: Request,
@@ -53,53 +56,10 @@ async def get_properties_feed(
 
 
 @router.get(
-    "/",
-    response_model=ManyPropertiesInResponse,
-    status_code=HTTPStatus.OK,
-)
-async def get_properties_by_user(
-    request: Request,
-    user: UserModel = Depends(get_current_user_authorizer()),
-    limit: int = Query(20, gt=0),
-    offset: int = Query(0, ge=0),
-    type: PropertyType = "",
-) -> Any:
-    """Get user properties"""
-    filters = PropertyFilterParams(
-        type=type, owner=user.username, limit=limit, offset=offset
-    )
-    properties = await get_user_properties(request, filters)
-    return create_aliased_response(
-        ManyPropertiesInResponse(
-            properties=properties, properties_count=len(properties)
-        )
-    )
-
-
-@router.post("/", response_model=PropertyInResponse, status_code=HTTPStatus.CREATED)
-async def new_property(
-    request: Request,
-    property: PropertyInCreate = Body(..., embed=True),
-    user: Optional[UserModel] = Depends(get_current_user_authorizer()),
-) -> Any:
-    """Create new property"""
-    if user:
-        if not user.is_active and not user.is_verified:
-            raise HTTPException(
-                status_code=HTTPStatus.BAD_REQUEST, detail="User not allowed"
-            )
-        new_property = await create_property(request, property, user)
-        if new_property:
-            return create_aliased_response(new_property)
-    raise HTTPException(
-        status_code=HTTPStatus.BAD_REQUEST, detail="Something went wrong / Bad request"
-    )
-
-
-@router.get(
-    "/{slug}",
+    "/feed/{slug}",
     status_code=HTTPStatus.OK,
     response_model=PropertyInResponse,
+    tags=["Properties"],
 )
 async def get_property(
     request: Request,
@@ -117,6 +77,51 @@ async def get_property(
 
 
 @router.get(
+    "/",
+    response_model=ManyPropertiesInResponse,
+    status_code=HTTPStatus.OK,
+)
+async def get_properties_by_user(
+    request: Request,
+    user: UserModel = Depends(get_current_user_authorizer()),
+    limit: int = Query(20, gt=0),
+    offset: int = Query(0, ge=0),
+    type: PropertyType = "",
+) -> Any:
+    """Get user properties"""
+    await check_user_permission(request, user)
+    filters = PropertyFilterParams(
+        type=type, owner=user.username, limit=limit, offset=offset
+    )
+    properties = await get_user_properties(request, filters)
+    return create_aliased_response(
+        ManyPropertiesInResponse(
+            properties=properties, properties_count=len(properties)
+        )
+    )
+
+
+@router.post(
+    "/",
+    response_model=PropertyInResponse,
+    status_code=HTTPStatus.CREATED,
+)
+async def new_property(
+    request: Request,
+    property: PropertyInCreate = Body(..., embed=True),
+    user: Optional[UserModel] = Depends(get_current_user_authorizer()),
+) -> Any:
+    """Create new property"""
+    await check_user_permission(request, user)
+    new_property = await create_property(request, property, user)
+    if new_property:
+        return create_aliased_response(new_property)
+    raise HTTPException(
+        status_code=HTTPStatus.BAD_REQUEST, detail="Something went wrong / Bad request"
+    )
+
+
+@router.get(
     "/{slug}",
     status_code=HTTPStatus.OK,
     response_model=PropertyInResponse,
@@ -127,17 +132,9 @@ async def get_user_property(
     user: Optional[UserModel] = Depends(get_current_user_authorizer()),
 ) -> Any:
     """Get user property by slug"""
+    await check_user_permission(request, user, slug)
+    await check_property_owner(request, slug, user.email)
     property = await get_property_by_slug(request, slug)
-    if not property:
-        raise HTTPException(
-            status_code=HTTPStatus.NOT_FOUND,
-            detail=f"Property with slug '{slug}' not found",
-        )
-    if not property.owner.username == user.username:
-        raise HTTPException(
-            status_code=HTTPStatus.UNAUTHORIZED,
-            detail=f"Property with slug '{slug}' not found",
-        )
     return create_aliased_response(PropertyInResponse(property=property))
 
 
@@ -153,6 +150,7 @@ async def update_property(
     user: Optional[UserModel] = Depends(get_current_user_authorizer()),
 ) -> Any:
     """Update Property"""
+    await check_user_permission(request, user, slug)
     await check_property_owner(request, slug, user.email)
     updated_property = await update_property_by_slug(request, slug, data)
     return create_aliased_response(PropertyInResponse(property=updated_property))
@@ -168,6 +166,29 @@ async def delete_property(
     user: Optional[UserModel] = Depends(get_current_user_authorizer()),
 ) -> Any:
     """Update Property"""
+    await check_user_permission(request, user, slug)
     await check_property_owner(request, slug, user.email)
     await delete_property_by_slug(request, slug, user.email)
     return {"message": "success"}
+
+
+@router.post(
+    "/{slug}/reservations",
+    response_model=ReservationModel,
+    status_code=HTTPStatus.CREATED,
+)
+async def new_reservation(
+    request: Request,
+    slug: str = Path(..., min_length=1),
+    reservation: ReservationModel = Body(..., embed=True),
+    user: Optional[UserModel] = Depends(get_current_user_authorizer()),
+):
+    """Create new reservation"""
+    await check_user_permission(request, user, slug)
+    await check_property_owner(request, slug, user.email)
+    new_reservation = await create_reservation(request, reservation, slug)
+    if new_reservation:
+        return create_aliased_response(ReservationModel(**new_reservation))
+    raise HTTPException(
+        status_code=HTTPStatus.BAD_REQUEST, detail="Something went wrong / Bad request"
+    )
